refactored utils to use a more modular approach
